Use `cp` instead of `ln` for tracking do files in prepped/
Due to linking it sometimes complains that I am writing to the file
to be tracked instead of a temporary file (because they are both linked to
the same inode on the filesystem).  Using `cp` now duplicates the file,
which isn't ideal, but better than getting errors during the run.  Would be nice
if the filesystem had copy-on-write...